+++ tests/basic.json
@@ -3,7 +3,8 @@
     {
       "name": "Alphabet",
       "cyr_taras": "Аа Бб Цц Цьць Чч Дд Дз дз Дзь дзь Дж дж Ээ Фф Гг Х х Іі Йй Кк Льль Лл Мм Нн Ньнь Оо Пп Рр Сс Сьсь Шш Тт Уу Ўў Вв Ыы Зз Зьзь Жж",
-      "lat": "Aa Bb Cc Ćć Čč Dd Dz dz Dź dź Dž dž Ee Ff Hh Ch ch Ii Jj Kk Ll Łł Mm Nn Ńń Oo Pp Rr Ss Śś Šš Tt Uu Ŭŭ Vv Yy Zz Źź Žž"
+      "lat": "Aa Bb Cc Ćć Čč Dd Dz dz Dź dź Dž dž Ee Ff Hh Ch ch Ii Jj Kk Ll Łł Mm Nn Ńń Oo Pp Rr Ss Śś Šš Tt Uu Ŭŭ Vv Yy Zz Źź Žž",
+      "lat_no_diactric": "Aa Bb Cc Cc Cc Dd Dz dz Dz dz Dz dz Ee Ff Hh Ch ch Ii Jj Kk Ll Ll Mm Nn Nn Oo Pp Rr Ss Ss Ss Tt Uu Uu Vv Yy Zz Zz Zz"
     },
     {
       "name": "Soft letters",
